Refactor unit tests to use objectProvider objects
