+++ src/test/kotlin/es/urjc/realfood/clients/application/CheckoutCartTest.kt
@@ -1,41 +1,12 @@
 package es.urjc.realfood.clients.application
 
-import es.urjc.realfood.clients.domain.*
-import es.urjc.realfood.clients.domain.repository.CartRepository
-import es.urjc.realfood.clients.domain.repository.OrderRepository
-import es.urjc.realfood.clients.domain.services.CheckoutCartService
+import es.urjc.realfood.clients.domain.CartObjectProvider.Companion.validCartItemDto
+import es.urjc.realfood.clients.domain.ClientObjectProvider.Companion.validClientId
+import es.urjc.realfood.clients.domain.OrderObjectProvider.Companion.validOrderId
 import es.urjc.realfood.clients.domain.services.CheckoutServiceRequest
 import es.urjc.realfood.clients.domain.services.CheckoutServiceResponse
-import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.TestInstance
-import org.mockito.Mockito.mock
-import org.springframework.boot.test.context.SpringBootTest
 
-@SpringBootTest(
-    classes = [
-        CartRepository::class,
-        OrderRepository::class
-    ]
-)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class CheckoutCartTest {
-
-    lateinit var cartRepository: CartRepository
-    lateinit var orderRepository: OrderRepository
-    lateinit var checkoutCartService: CheckoutCartService
-    lateinit var checkoutCart: CheckoutCart
-
-    @BeforeAll
-    fun init() {
-        cartRepository = mock(CartRepository::class.java)
-        orderRepository = mock(OrderRepository::class.java)
-        checkoutCartService = mock(CheckoutCartService::class.java)
-        checkoutCart = CheckoutCart(
-            cartRepository = cartRepository,
-            orderRepository = orderRepository,
-            checkoutCartService = checkoutCartService
-        )
-    }
+abstract class CheckoutCartTest {
 
     protected fun validCheckoutCartRequest(): CheckoutCartRequest {
         return CheckoutCartRequest(
@@ -69,53 +40,5 @@ class CheckoutCartTest {
             orderId = null
         )
     }
-
-    protected fun validCart(): Cart {
-        val cart = Cart(
-            id = validCartId(),
-            client = validClient()
-        )
-        cart.items[validItemId()] = validCartItem()
-        return cart
-    }
-
-    protected fun invalidCart(): Cart {
-        return Cart(
-            id = validCartId(),
-            client = validClient()
-        )
-    }
-
-    protected fun validClient(): Client {
-        return Client(
-            id = validClientId(),
-            name = Name("Cristofer"),
-            lastName = LastName("Lopez"),
-            email = Email("cristofer@cristofer.es"),
-            password = Password("1234")
-        )
-    }
-
-    protected fun validCartItem(): CartItem {
-        return CartItem(
-            itemId = validItemId(),
-            quantity = 5
-        )
-    }
-
-    protected fun validCartItemDto(): es.urjc.realfood.clients.domain.services.CartItemDto {
-        return es.urjc.realfood.clients.domain.services.CartItemDto(
-            item = validItemId(),
-            quantity = 5
-        )
-    }
-
-    protected fun validCartId(): CartId = CartId("89a135b8-98dc-4e57-a22f-b5f99c6b1a99")
-
-    protected fun validClientId(): ClientId = ClientId("89a135b8-98dc-4e57-a22f-b5f99c6b1a00")
-
-    protected fun validOrderId(): OrderId = OrderId("89a135b8-98dc-4e57-a22f-b5f99c6b1a11")
-
-    protected fun validItemId(): String = "89a135b8-98dc-4e57-a22f-b5f99c6b1a11"
 
 }
