+++ src/test/kotlin/es/urjc/realfood/clients/application/FindByIdClientTest.kt
@@ -1,49 +1,18 @@
 package es.urjc.realfood.clients.application
 
-import es.urjc.realfood.clients.domain.*
-import es.urjc.realfood.clients.domain.repository.ClientRepository
-import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.TestInstance
-import org.mockito.Mockito
-import org.springframework.boot.test.context.SpringBootTest
+import es.urjc.realfood.clients.domain.ClientObjectProvider.Companion.validClientIdString
 
-@SpringBootTest(
-    classes = [
-        ClientRepository::class
-    ]
-)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 abstract class FindByIdClientTest {
-
-    lateinit var clientRepository: ClientRepository
-    lateinit var findByIdClient: FindByIdClient
-
-    @BeforeAll
-    fun init() {
-        clientRepository = Mockito.mock(ClientRepository::class.java)
-
-        findByIdClient = FindByIdClient(
-            clientRepository = clientRepository
-        )
-    }
-
-    protected fun validClient(): Client {
-        return Client(
-            id = ClientId("89a135b8-98dc-4e57-a22f-b5f99c6b1a99"),
-            name = Name("Cristofer"),
-            lastName = LastName("Lopez"),
-            email = Email("cristofer@cristofer.es"),
-            password = Password("1234")
-        )
-    }
-
-    protected fun validClientId(): ClientId {
-        return ClientId("89a135b8-98dc-4e57-a22f-b5f99c6b1a99")
-    }
 
     protected fun validFindByIdClientRequest(): FindByIdClientRequest {
         return FindByIdClientRequest(
-            id = "89a135b8-98dc-4e57-a22f-b5f99c6b1a99"
+            id = validClientIdString()
+        )
+    }
+
+    protected fun invalidFindByIdClientRequest(): FindByIdClientRequest {
+        return FindByIdClientRequest(
+            id = "INVALID-ID"
         )
     }
