+++ src/test/kotlin/es/urjc/realfood/clients/unit/DeleteClientUnitaryTest.kt
@@ -1,13 +1,39 @@
 package es.urjc.realfood.clients.unit
 
+import es.urjc.realfood.clients.application.DeleteClient
 import es.urjc.realfood.clients.application.DeleteClientRequest
 import es.urjc.realfood.clients.application.DeleteClientTest
+import es.urjc.realfood.clients.domain.ClientObjectProvider.Companion.validClient
+import es.urjc.realfood.clients.domain.ClientObjectProvider.Companion.validClientId
 import es.urjc.realfood.clients.domain.exception.EntityNotFoundException
+import es.urjc.realfood.clients.domain.repository.ClientRepository
 import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
+import org.mockito.Mockito
 import org.mockito.Mockito.*
+import org.springframework.boot.test.context.SpringBootTest
 
+@SpringBootTest(
+    classes = [
+        ClientRepository::class
+    ]
+)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class DeleteClientUnitaryTest : DeleteClientTest() {
+
+    lateinit var clientRepository: ClientRepository
+    lateinit var deleteClient: DeleteClient
+
+    @BeforeAll
+    fun init() {
+        clientRepository = mock(ClientRepository::class.java)
+
+        deleteClient = DeleteClient(
+            clientRepository = clientRepository
+        )
+    }
 
     @Test
     fun `given valid id when delete user then ok`() {
