+++ src/test/kotlin/es/urjc/realfood/clients/unit/FindByIdAndClientIdOrderUnitaryTest.kt
@@ -1,13 +1,40 @@
 package es.urjc.realfood.clients.unit
 
+import es.urjc.realfood.clients.application.FindByIdAndClientIdOrder
 import es.urjc.realfood.clients.application.FindByIdAndClientIdOrderTest
+import es.urjc.realfood.clients.domain.ClientObjectProvider.Companion.validClientId
+import es.urjc.realfood.clients.domain.OrderObjectProvider.Companion.validOrder
+import es.urjc.realfood.clients.domain.OrderObjectProvider.Companion.validOrderId
 import es.urjc.realfood.clients.domain.exception.EntityNotFoundException
-import org.junit.jupiter.api.Assertions
-import org.junit.jupiter.api.Assertions.assertEquals
+import es.urjc.realfood.clients.domain.repository.CartRepository
+import es.urjc.realfood.clients.domain.repository.OrderRepository
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
+import org.mockito.Mockito
 import org.mockito.Mockito.`when`
+import org.springframework.boot.test.context.SpringBootTest
 
+@SpringBootTest(
+    classes = [
+        CartRepository::class,
+        OrderRepository::class
+    ]
+)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class FindByIdAndClientIdOrderUnitaryTest : FindByIdAndClientIdOrderTest() {
+
+    lateinit var orderRepository: OrderRepository
+    lateinit var findByIdAndClientIdOrder: FindByIdAndClientIdOrder
+
+    @BeforeAll
+    fun init() {
+        orderRepository = Mockito.mock(OrderRepository::class.java)
+        findByIdAndClientIdOrder = FindByIdAndClientIdOrder(
+            orderRepository = orderRepository,
+        )
+    }
 
     @Test
     fun `given valid request when get user order then return order`() {
@@ -22,20 +49,20 @@ class FindByIdAndClientIdOrderUnitaryTest : FindByIdAndClientIdOrderTest() {
 
     @Test
     fun `given invalid client id when get user order then return illegal argument exception`() {
-        val exc = Assertions.assertThrows(IllegalArgumentException::class.java) {
+        val exc = assertThrows(IllegalArgumentException::class.java) {
             findByIdAndClientIdOrder(invalidClientIdRequest())
         }
 
-        Assertions.assertTrue(exc.message!!.contains("Invalid UUID"))
+        assertTrue(exc.message!!.contains("Invalid UUID"))
     }
 
     @Test
     fun `given invalid order id when get user order then return illegal argument exception`() {
-        val exc = Assertions.assertThrows(IllegalArgumentException::class.java) {
+        val exc = assertThrows(IllegalArgumentException::class.java) {
             findByIdAndClientIdOrder(invalidOrderIdRequest())
         }
 
-        Assertions.assertTrue(exc.message!!.contains("Invalid UUID"))
+        assertTrue(exc.message!!.contains("Invalid UUID"))
     }
 
     @Test
@@ -43,7 +70,7 @@ class FindByIdAndClientIdOrderUnitaryTest : FindByIdAndClientIdOrderTest() {
         `when`(orderRepository.findByIdAndClientId(validOrderId(), validClientId()))
             .thenReturn(null)
 
-        val exc = Assertions.assertThrows(EntityNotFoundException::class.java) {
+        val exc = assertThrows(EntityNotFoundException::class.java) {
             findByIdAndClientIdOrder(validRequest())
         }
